requirements.txt
opencv-python 
matplotlib 
ipykernel
fastapi
pillow
mediapipe
imutils
uvicorn
python-multipart
# 192.168.0.106